Changed fabs() to fabsf().

+++ src/PIDF.cpp
@@ -28,7 +28,7 @@ float PIDF::updateDelta(float measurement, float measurementDelta, float deltaT)
 {
     const float error = _setpoint - measurement;
 
-    if (fabs(error) > _integralThreshold) {
+    if (fabsf(error) > _integralThreshold) {
         // "integrate" the error
         //_errorIntegral += _pid.ki*_error*deltaT; // Euler integration
         _errorIntegral += _pid.ki*0.5F*(error + _errorPrevious)*deltaT; // integration using trapezoid rule
@@ -38,7 +38,7 @@ float PIDF::updateDelta(float measurement, float measurementDelta, float deltaT)
         }
     }
     const float pValue = _pid.kp*error;
-    if (_outputSaturationValue != 0.0F && fabs(pValue) > _outputSaturationValue) {
+    if (_outputSaturationValue != 0.0F && fabsf(pValue) > _outputSaturationValue) {
         // The output is saturated by the pValue, so the integral value will not increase output and will result in overshoot when the pValue eventually comes down,
         // so set the integral to zero.
         _errorIntegral = 0.0F;
